Moving code around a little bit
