+++ kuristo/_utils.py
@@ -1,9 +1,6 @@
 import sys
 import subprocess
 import os
-import yaml
-from .scanner import Scanner
-from .job_spec import JobSpec
 from jinja2 import Template
 
 
@@ -20,37 +17,6 @@ def get_default_core_limit():
         except Exception:
             pass  # fallback below
     return os.cpu_count() or 1
-
-
-def scan_locations(locations):
-    """
-    Scan the locations for the workflow files
-    """
-    workflow_files = []
-    for loc in locations:
-        scanner = Scanner(loc)
-        workflow_files.extend(scanner.scan())
-    return workflow_files
-
-
-def specs_from_file(file_path):
-    specs = []
-    with open(file_path, 'r') as file:
-        data = yaml.safe_load(file)
-        jobs = data.get('jobs', {})
-        for t, params in jobs.items():
-            specs.append(JobSpec.from_dict(t, params))
-    return specs
-
-
-def parse_workflow_files(workflow_files):
-    """
-    Parse workflow files (ktests.yaml)
-    """
-    specs = []
-    for file in workflow_files:
-        specs.extend(specs_from_file(file))
-    return specs
 
 
 def resolve_path(path_str, source_root, build_root):
